File system chunk storage implemented. Purging of directory is performed before start.
